backend/distribution/api/v1/serializers.py
from rest_framework import serializers

from distribution.models import Distribution


class DistributionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Distribution
        fields = '__all__'
